% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunction.R
\name{AnalysisFunction}
\alias{AnalysisFunction}
\title{AnalysisFunction Function}
\usage{
AnalysisFunction(x, training)
}
\arguments{
\item{x}{vector of numbers}

\item{training}{boolean value. If T, random noise is added. If F,
no noise is added, vector is returned.}
}
\value{
vector of numbers
}
\description{
This function takes a vector of numbers and adds random noise to each entry
taken from a uniform distribution (-1, 1) if training is set to True. If set
to False, it returns the same vector.
}
\examples{
AnalysisFunction(x, T)
AnalysisFunction(x, F)
}
