% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataAsset.R
\name{GetDataAsset}
\alias{GetDataAsset}
\title{GetDataAsset Function}
\usage{
GetDataAsset()
}
\arguments{
\item{no}{input}
}
\value{
data table
}
\description{
This function downloads a .zip file from Dropbox and imports it into R
as a data table. It deletes the files on disk after it is imported.
}
\examples{
GetDataAsset()
}
